requirements.txt
coincurve~=19.0.0
green~=4.0.2
protobuf~=3.19.1
pycryptodome~=3.19.1
freesimplegui
python-dotenv
